Change servlet to controller
