+++ projects/src/main/java/timeboard/projects/ProjectsConfiguration.java
@@ -26,12 +26,10 @@ package timeboard.projects;
  * #L%
  */
 
-import org.springframework.boot.web.servlet.ServletComponentScan;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 @ComponentScan({"timeboard.projects"})
-@ServletComponentScan({"timeboard.projects"})
 public class ProjectsConfiguration {
 }
