+++ timesheet/src/main/java/timeboard/timesheet/TimesheetConfiguration.java
@@ -26,12 +26,10 @@ package timeboard.timesheet;
  * #L%
  */
 
-import org.springframework.boot.web.servlet.ServletComponentScan;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 @ComponentScan("timeboard.timesheet")
-@ServletComponentScan("timeboard.timesheet")
 public class TimesheetConfiguration {
 }
