+++ timesheet/src/main/java/timeboard/timesheet/TimesheetController.java
@@ -55,7 +55,7 @@ import java.util.List;
 
 @Controller
 @RequestMapping("/org/{orgID}/timesheet")
-public class TimesheetServlet {
+public class TimesheetController {
 
     private static final DateFormat DATE_FORMAT = new SimpleDateFormat("yyyy-MM-dd");
     private static final ObjectMapper MAPPER = new ObjectMapper();
